+++ src/BLE_Client.cpp
@@ -31,7 +31,7 @@ void SpinBLEClient::start() {
   // Create the task for the BLE Client loop
   xTaskCreatePinnedToCore(bleClientTask,   /* Task function. */
                           "BLEClientTask", /* name of task. */
-                          5500,            /* Stack size of task */
+                          BLE_CLIENT_STACK,            /* Stack size of task */
                           NULL,            /* parameter of the task */
                           1,               /* priority of the task  */
                           &BLEClientTask,  /* Task handle to keep track of created task */
@@ -557,6 +557,10 @@ void SpinBLEClient::resetDevices(NimBLEClient *pClient) {
 // Control a connected FTMS trainer. If no args are passed, treat it like an external stepper motor.
 void SpinBLEClient::FTMSControlPointWrite(const uint8_t *pData, int length) {
   NimBLEClient *pClient = nullptr;
+  uint8_t modData[7];
+  for (int i = 0; i < length; i++) {
+    modData[i] = pData[i];
+  }
   for (int i = 0; i < NUM_BLE_DEVICES; i++) {
     if (myBLEDevices[i].postConnected && (myBLEDevices[i].serviceUUID == FITNESSMACHINESERVICE_UUID)) {
       if (NimBLEDevice::getClientByPeerAddress(myBLEDevices[i].peerAddress)->getService(FITNESSMACHINESERVICE_UUID)) {
@@ -567,9 +571,23 @@ void SpinBLEClient::FTMSControlPointWrite(const uint8_t *pData, int length) {
   }
   if (pClient) {
     NimBLERemoteCharacteristic *writeCharacteristic = pClient->getService(FITNESSMACHINESERVICE_UUID)->getCharacteristic(FITNESSMACHINECONTROLPOINT_UUID);
+    int logBufLength                                = 0;
     if (writeCharacteristic) {
-      writeCharacteristic->writeValue(pData, length);
-      SS2K_LOG(BLE_CLIENT_LOG_TAG, "Sent FTMS");
+      const int kLogBufCapacity = length + 40;
+      char logBuf[kLogBufCapacity];
+      if (modData[0] == FitnessMachineControlPointProcedure::SetIndoorBikeSimulationParameters) {  // use virtual Shifting
+        int incline = ss2k.targetPosition / userConfig.getInclineMultiplier();
+        modData[3]  = (uint8_t)(incline & 0xff);
+        modData[4]  = (uint8_t)(incline >> 8);
+        writeCharacteristic->writeValue(modData, length);
+        logBufLength = ss2k_log_hex_to_buffer(modData, length, logBuf, 0, kLogBufCapacity);
+        logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Shifted Sim Data: %d", rtConfig.getShifterPosition());
+      } else {
+        writeCharacteristic->writeValue(modData, length);
+        logBufLength = ss2k_log_hex_to_buffer(modData, length, logBuf, 0, kLogBufCapacity);
+        logBufLength += snprintf(logBuf + logBufLength, kLogBufCapacity - logBufLength, "-> Shifted ERG Data: %d", rtConfig.getShifterPosition());
+      }
+      SS2K_LOG(BLE_CLIENT_LOG_TAG, "%s", logBuf);
     }
   }
 }